backend/pyproject.toml
[project]
name = "chiron-backend"
version = "0.1.0"
description = "Chiron Test — FastAPI backend (Phase 0 Foundation)"
authors = [
    {name = "Chiron Team"}
]
readme = "README.md"
requires-python = "^3.11"
dependencies = [
    "fastapi (>=0.135.1,<0.136.0)",
    "uvicorn[standard] (>=0.41.0,<0.42.0)",
    "sqlalchemy (>=2.0.48,<3.0.0)",
    "psycopg2-binary (>=2.9.11,<3.0.0)",
    "pydantic-settings (>=2.13.1,<3.0.0)",
    "alembic (>=1.18.4,<2.0.0)",
    "email-validator (>=2.0.0)",
    "python-multipart (>=0.0.12)"
]

[project.optional-dependencies]
dev = [
    "pytest>=9.0.0",
    "pytest-cov>=7.0.0",
    "pytest-mock>=3.15.0",
    "httpx>=0.28.0",
    "anyio>=4.0.0",
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "strict"
markers = [
    "integration: marks tests as integration tests requiring a live PostgreSQL database (deselect with '-m \"not integration\"')",
]
addopts = "-m 'not integration'"
